tests 01-04 test_dash_fns
