+++ apps/web/pages/workout_log.py
@@ -4,8 +4,8 @@ import pandas as pd
 import requests
 from typing import List
 from constants import ROOT_URL
-from apps.web.dash_fxs import flask_requests_get, flask_requests_post
-from apps.web.dash_fxs import get_name, seconds_to_duration
+from apps.web.dash_fns import flask_requests_get, flask_requests_post
+from apps.web.dash_fns import get_name, seconds_to_duration
 import dash_bootstrap_components as dbc
 import plotly.express as px
 from dash.exceptions import PreventUpdate
